Remove support for 2.10.x and 2.13.0-M4.
