+++ .travis.yml
@@ -6,7 +6,6 @@ install:
   - pip install --user codecov
 
 scala:
-- 2.10.7
 - 2.11.12
 - 2.12.10
 - 2.13.1
