+++ core/src/test/scala/shapeless/labelledgeneric.scala
@@ -79,6 +79,13 @@ object LabelledGenericTestsAux {
   }
 }
 
+object ShapelessTaggedAux {
+  import tag.@@
+
+  trait CustomTag
+  case class Dummy(i: Int @@ CustomTag)
+}
+
 object ScalazTaggedAux {
   import labelled.FieldType
 
@@ -384,6 +391,16 @@ class LabelledGenericTests {
   }
 
   @Test
+  def testShapelessTagged: Unit = {
+    import ShapelessTaggedAux._
+
+    val lgen = LabelledGeneric[Dummy]
+    val s = s"${lgen from Record(i=tag[CustomTag](0))}"
+    assertEquals(s, "Dummy(0)")
+  }
+
+
+  @Test
   def testScalazTagged: Unit = {
     import ScalazTaggedAux._
 
@@ -397,6 +414,14 @@ class LabelledGenericTests {
 
     implicitly[TC[DummyTagged]]
 
-    // Note: Further tests in LabelledGeneric211Tests
+    type R = Record.`'i -> Int @@ CustomTag`.T
+    val lgen = LabelledGeneric[Dummy]
+    implicitly[lgen.Repr =:= R]
+    implicitly[TC[R]]
+
+    type RT = Record.`'b -> Boolean, 'i -> Int @@ CustomTag`.T
+    val lgent = LabelledGeneric[DummyTagged]
+    implicitly[lgent.Repr =:= RT]
+    implicitly[TC[RT]]
   }
 }
