--- core/src/test/scala_2.11+/shapeless/implicits.scala
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2015 Miles Sabin
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package shapeless
-
-import org.junit.Test
-import org.junit.Assert._
-
-import test.illTyped
-
-class Cached211Test {
-  trait Foo[A]
-  object Foo {
-    implicit def materialize[A]: Foo[A] = new Foo[A] {}
-  }
-
-  @Test
-  def testAmbiguous: Unit = {
-    implicit val a = new Foo[String] { }
-    implicit val b = new Foo[String] { }
-    illTyped(
-      "cachedImplicit[Foo[String]]"
-    )
-  }
-}
